Added a basic probabilistic method for choosing items
Should work relatively well, no item should be displayed significantly fewer times than any other items, although it is possible in theory. There is still a problem with the same item appearing twice in a row. A new weighting method is required to take this into account, and may also need to keep track of the last time a specific item appeared.